requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
pydantic>=2.0
playwright>=1.52.0
beautifulsoup4>=4.12.0
google-adk>=0.5.0
google-cloud-aiplatform>=1.74.0
termcolor>=3.1.0
rich>=13.0
python-dotenv>=1.0.0
